add support for windows ssh-agent
Windows 10 has the optional feature for an ssh client.
This package comes with a backed in ssh-agent which
utilizes windows named pipes instead of unix sockets.
This CL adds support to communicate which such ssh-agent.
This is mainly needed to enable ssh agents which do
not support the pageant established "protocol".

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>