README.md: Remove and replace sockets when they exist.
Currently, if gpg-agent/ssh-agent have already started (which they seem to have in my wsl2), then this just exits immediately because grep returns 0, and doesn't re-map. I'm not sure if this was an attempt to prevent multiple wsl2-ssh-pageant.exe instances from clobbering each other, but it seems like given your first action is an rm-rf, it should have a socket there already to delete/replace.